refactor(app/globals.css): modularized independent css classes, using glabals for overrides
example of override: the footer daisyui class media query override. currently resorting to placing
in globals.css to avoid additional depdencies for classname resolution

complete #12 , resolve and close #11
