+++ app/(client-app)/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingSpinner } from '@/components/LoadingSpinner';
+import { LoadingSpinner } from '@/components/loading/LoadingSpinner';
 import withCenteredWrapper from '@/hocs/withCenteredWrapper';
 
 export default withCenteredWrapper(LoadingSpinner);
