+++ app/test/page.tsx
@@ -2,8 +2,8 @@
 
 import type { JSX } from 'react';
 
-import { LoadingSpinner } from '@/components/LoadingSpinner';
 import FatalError from '@/components/error/FatalError';
+import { LoadingSpinner } from '@/components/loading/LoadingSpinner';
 import withCenteredWrapper from '@/hocs/withCenteredWrapper';
 
 export default function TestPage(): JSX.Element {
@@ -11,6 +11,6 @@ export default function TestPage(): JSX.Element {
   const WrappedLoadingSpinner = withCenteredWrapper(LoadingSpinner);
   const WrappedFatalError = withCenteredWrapper(FatalError);
 
-  // return <WrappedLoadingSpinner />;
-  return <WrappedFatalError />;
+  return <WrappedLoadingSpinner />;
+  // return <WrappedFatalError />;
 }
