+++ components/loading/LoadingSpinner.module.css
@@ -1,0 +1,12 @@
+.loading-spinner-image {
+  animation: spin-image 3s linear infinite;
+}
+
+@keyframes spin-image {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
